refactor: 이미지 경로 수정
